Move deploy method to config class
